a6plte: configs: media: remove unsupported media codecs and profiles
Only the qualities specified inside
frameworks/av/media/libmedia/MediaProfiles.cpp
are supported.
Remvoe unsupported ones.

03-05 17:46:16.749   977   977 E MediaCodecsXmlParser: Cannot find the
role for a decoder of type video/divx
03-05 17:46:16.749   977   977 E MediaCodecsXmlParser: Cannot find the
role for a decoder of type video/divx311
03-05 17:46:16.749   977   977 E MediaCodecsXmlParser: Cannot find the
role for a decoder of type video/divx4
03-05 17:46:16.749   977   977 E MediaCodecsXmlParser: Cannot find the
role for a decoder of type video/x-ms-wmv
03-05 17:46:16.749   977   977 E MediaCodecsXmlParser: Cannot find the
role for a decoder of type video/x-ms-wmv

Change-Id: Id6a32df48d87c8f530e76e763a6aae8908fc232d